decision-trainer-py/pyproject.toml
[project]
name = "decision_trainer"
version = "0.1.0"
description = "decision_trainer using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=0.141.0,<1.0.0",
    "python-dotenv>=0.21.0",
    "weave>=0.51.0",
    "openai>=1.0.0",
    "crewai-tools[mcp]>=0.51.1",
    "requests>=2.32.4",
    "pydantic>=2.11.7,<2.12",
]

[project.scripts]
decision_trainer = "decision_trainer.main:run"
run_crew = "decision_trainer.main:run"
train = "decision_trainer.main:train"
replay = "decision_trainer.main:replay"
test = "decision_trainer.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
